+ variant with reuse of transposed chuncks of matrix B
Transposed chunks of matrix B, coresponding to 768x150 rectangles, are saved and reused
over full range of M. Added outermost level of blocking - n_superstep=768
This variant is intended for a case of N=integer multiple of 4096 that does not work well on Haswell.
I am not sutisfied with result.
Although it does handle a bad case ~10% faster than original 5x2 variant, on the absolute scale it is still up to 11% slower than original.