+ns3x3
Add ns3x3 - a variant with inner loop that multiplies 3 rows of A by 3x8 columns of B.
By itself this variant is not intersting - even on Kabi Lake it is 7-15% slower than ns4x3.
I suppose, on Haswell/Broadwell the dufference is even bigger.
It is impleented as a step toward ns3x4 and then ns2x5.